+++ release_notes.md
@@ -2,3 +2,4 @@
 
 ### 1.0.0
 - Initial version
+- Added `getDefaultHost` and `getDefaultPort` functions to `rcc-ioredis-adapter` module